@@ -9,30 +9,26 @@
   "author": "DucHoang",
   "license": "ISC",
   "dependencies": {
-    "binance": "^2.13.18",
-    "csv-parse": "^5.6.0",
-    "grammy": "^1.34.0",
-    "lightweight-charts": "^5.0.8",
-    "luxon": "^3.5.0",
-    "moment": "^2.30.1",
-    "papaparse": "^5.5.3"
-  },
-  "devDependencies": {
-    "commander": "^2.20.3",
-    "cron": "^3.5.0",
-    "@babel/core": "^7.27.4",
+    "@babel/core": "^7.28.0",
     "@babel/plugin-proposal-optional-chaining": "^7.21.0",
-    "@babel/plugin-transform-runtime": "^7.27.4",
-    "@babel/preset-env": "^7.0.0-beta.51",
-    "autoprefixer": "^9.0.0",
-    "babel-loader": "^8.3.0",
-    "css-loader": "^3.6.0",
-    "extract-text-webpack-plugin": "^4.0.0-beta.0",
-    "mini-css-extract-plugin": "^0.9.0",
-    "postcss-loader": "^4.3.0",
-    "sass-loader": "^10.5.2",
-    "webpack": "^4.46.0",
-    "webpack-cli": "^4.7.2",
-    "shelljs": "^0.8.5"
+    "@babel/plugin-transform-runtime": "^7.28.0",
+    "@babel/preset-env": "^7.28.0",
+    "autoprefixer": "^10.4.21",
+    "babel-loader": "^10.0.0",
+    "commander": "^14.0.0",
+    "cron": "^4.3.2",
+    "css-loader": "^7.1.2",
+    "csv-parse": "^6.0.0",
+    "grammy": "^1.37.0",
+    "lightweight-charts": "^5.0.8",
+    "luxon": "^3.7.1",
+    "mini-css-extract-plugin": "^2.9.2",
+    "moment": "^2.30.1",
+    "papaparse": "^5.5.3",
+    "postcss-loader": "^8.1.1",
+    "sass-loader": "^16.0.5",
+    "shelljs": "^0.10.0",
+    "webpack": "^5.100.1",
+    "webpack-cli": "^6.0.1"
   }
 }
